+++ src/flythings_dataspace_sdk/connector/clients/__init__.py
@@ -1,3 +1,4 @@
+from flythings_dataspace_sdk.connector.clients.assets import AssetsClient
 from flythings_dataspace_sdk.connector.clients.contract_agreements import ContractAgreementsClient
 from flythings_dataspace_sdk.connector.clients.edrs import EDRSClient
 from flythings_dataspace_sdk.connector.clients.transfers import TransfersClient